Add leading slashes to js and css
Without the leading slash _can_ work, depending on how the domain is
setup and where the files all live, but the leading slash is more
explicit.

Move the script tag to right before the closing `body` tag. This is the
convention as it makes sure all the HTML is loaded before the script is
fired.